+++ lib/docking.dart
@@ -2,4 +2,5 @@ library docking;
 
 export 'src/layout/docking_layout.dart';
 export 'src/on_item_selection.dart';
+export 'src/on_item_close.dart';
 export 'src/docking.dart';